pyproject.toml
[project]
name = "frls-assignment"
version = "0.1.0"
description = "Baseline experiments for document matching across delivery notes and purchase invoices"
readme = "baseline_vector/README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0.0",
    "openai>=1.30.0",
    "pandas>=2.2.0",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "pypdf>=5.0.0",
    "python-dotenv>=1.0.1",
    "rapidfuzz>=3.9.0",
    "scikit-learn>=1.5.0",
    "sentence-transformers>=3.0.0",
    "tabulate>=0.9.0",
    "tqdm>=4.66.0",
    "torch>=2.3.0",
    "python-dateutil>=2.9.0.post0",
]

[tool.uv]
package = false
